requirements.txt
streamlit
langchain-core
langchain-groq
python-dotenv
torch

